Split settings to separate components
